register relative paths when importing tiltseries mdoc import: fallback in form params for magnification, sampling rate and voltage Use ImageFile from mdoc, falling back on ".mrcs" Start id of tiltimages in 1 Tests refactored: split in 2 files Test for links and relative paths
